one component for profile edit or create
